rewrote dsp_soundtouch.cpp

+++ foo_dsp_effect/SoundTouch/InterpolateCubic.h
@@ -8,7 +8,7 @@
 ///
 ////////////////////////////////////////////////////////////////////////////////
 //
-// $Id: InterpolateCubic.h 225 2015-07-26 14:45:48Z oparviai $
+// $Id: InterpolateCubic.h 179 2014-01-06 18:41:42Z oparviai $
 //
 ////////////////////////////////////////////////////////////////////////////////
 //
@@ -56,7 +56,7 @@ protected:
                         const SAMPLETYPE *src, 
                         int &srcSamples);
 
-    double fract;
+    float fract;
 
 public:
     InterpolateCubic();
