+++ foo_dsp_effect/SoundTouch/InterpolateShannon.h
@@ -13,7 +13,7 @@
 ///
 ////////////////////////////////////////////////////////////////////////////////
 //
-// $Id: InterpolateShannon.h 225 2015-07-26 14:45:48Z oparviai $
+// $Id: InterpolateShannon.h 179 2014-01-06 18:41:42Z oparviai $
 //
 ////////////////////////////////////////////////////////////////////////////////
 //
@@ -61,7 +61,7 @@ protected:
                         const SAMPLETYPE *src, 
                         int &srcSamples);
 
-    double fract;
+    float fract;
 
 public:
     InterpolateShannon();
